Build and debug user show and story forms

+++ app/controllers/api/stories_controller.rb
@@ -1,7 +1,10 @@
 class Api::StoriesController < ApplicationController
+  before_action :require_signed_in?, only: [:create, :update, :destroy]
+
   def create
     @story = Story.new(story_params)
     @story.img_url = "story_default.png" if @story.img_url == ""
+    @story.author_id = current_user.id
 
     if @story.save
       render :show
@@ -13,11 +16,15 @@ class Api::StoriesController < ApplicationController
   def update
     @story = Story.find(params[:id])
 
-    if @story.update(story_params)
-      @story.img_url = "story_default.png" if @story.img_url == ""
-      render :show
+    if @story.author_id == current_user.id
+      if @story.update(story_params)
+        @story.img_url = "story_default.png" if @story.img_url == ""
+        render :show
+      else
+        render json: @story.errors.full_messages, status: 422
+      end
     else
-      render json: @story.errors.full_messages, status: 422
+      render json: ["Please don't edit posts you haven't authored."], status: 422
     end
   end
 
@@ -31,18 +38,21 @@ class Api::StoriesController < ApplicationController
 
   def destroy
     @story = Story.find(params[:id])
-
-    if @story.destroy
-      render :show
+    
+    if @story.author_id == current_user.id
+      if @story.destroy
+        render :show
+      else
+        render json: @story.errors.full_messages, status: 422
+      end
     else
-      render json: @story.errors.full_messages, status: 422
+      render json: ["Please don't delete posts you haven't authored."], status: 422
     end
   end
 
   private
 
   def story_params
-    # TODO Remove :author_id after testing...
-    params.require(:story).permit(:title, :body, :img_url, :author_id)
+    params.require(:story).permit(:title, :body, :img_url)
   end
 end
